READY FOR FIGHT

+++ database/migrations/2024_01_13_130214_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('image');
             $table->string('nama_produk');
             $table->integer('stok');
-            $table->integer('harga');
+            $table->decimal('harga', 10, 3);
             $table->timestamps();            
         });
     }
