+++ database/migrations/2024_01_31_010941_create_keranjangs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('kode_produk');
             $table->integer('qty');
-            $table->integer('subtotal');
+            $table->decimal('subtotal', 10, 3);
             $table->timestamps();
 
             $table->foreign('kode_produk')->references('kode_produk')->on('products');
