+++ database/migrations/2024_01_31_014327_create_detail_transaksis_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('kode_produk');
             $table->unsignedBigInteger('no_invoice');
             $table->integer('qty');
-            $table->integer('subtotal');
+            $table->decimal('subtotal', 10, 3);
             $table->timestamps();
             $table->foreign('kode_produk')->references('kode_produk')->on('products');
             $table->foreign('no_invoice')->references('no_invoice')->on('transaksis');
